Update pattern matching implementation: support region searches & other usages
`findPattern` implementation updated, new usages:
- findPattern(handle, pattern, flags, patternOffset);
  => searches all regions & modules
- findPattern(handle, moduleName, pattern, flags, patternOffset);
  => searches a specified module
- findPattern(handle, baseAddress, pattern, flags, patternOffset);
  => searches only the module or region where the given base
     address lies inside

The `signatureOffset` parameter has been removed to simplify the function.
The value of the parameter was added to the returned address, this
can be applied outside of the function.

This change allows for pattern matching inside of regions, and not just
modules. It also allows users to do a blanket search of the process
if a module name or base address is not specified.